+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * (crypto/keyring) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Fix error wrapping in pkg crypto/keyring. Removed `cockroachdb/errors` in favour of `errorsmod.Wrap`.
 * (x/authz) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Added a limit of 200 grants pruned per `BeginBlock` and the `PruneExpiredGrants` message that prunes 75 expired grants on every run.
 * (x/consensus) [#1](https://github.com/atomone-hub/cosmos-sdk/pull/1) Address bug where sending `MsgUpdateParams` to `x/consensus` would overwrite `AppVersion` with zero value. Always use existing `AppVersion` param.
+* (x/epochs) [#33](https://github.com/atomone-hub/cosmos-sdk/pull/33) Add `x/epochs` module.
 
 ### Improvements